requirements.txt
streamlit
langchain
langchain-openai
langchain-community
langchain-core
chromadb
unstructured
unstructured-client
pypdf
python-dotenv
unstructured-loader